+++ docker/root/fonts.install.d/20-sourceserifpro.sh
@@ -1,0 +1,8 @@
+# -*- mode: sh; coding: utf-8-unix; -*-
+# Name: Adobe Source Serif Pro
+# License: SIL OFL 1.1
+# License URL: https://github.com/adobe-fonts/source-serif-pro/blob/master/LICENSE.md
+curl -L -O https://github.com/adobe-fonts/source-serif-pro/archive/2.007R-ro/1.007R-it.tar.gz
+tar xzf 1.007R-it.tar.gz
+mv source-serif-pro-2.007R-ro-1.007R-it/OTF /usr/share/fonts/opentype/sourceserifpro
+mv source-serif-pro-2.007R-ro-1.007R-it/LICENSE.txt /usr/share/fonts/opentype/sourceserifpro